requirements.txt
streamlit
pandas
openpyxl
pyyaml
scikit-learn
